+++ examples/json_serializable/test/fixture/json_serializable.auto_mappr.dart
@@ -5,9 +5,9 @@
 // AutoMapprGenerator
 // **************************************************************************
 
-// ignore_for_file: type=lint, unnecessary_cast, unused_local_variable
+// ignore_for_file: no_leading_underscores_for_library_prefixes, type=lint
+// ignore_for_file: unnecessary_cast, unused_local_variable
 
-// ignore_for_file: no_leading_underscores_for_library_prefixes
 import 'package:auto_mappr_annotation/auto_mappr_annotation.dart' as _i1;
 
 import 'json_serializable.dart' as _i2;
@@ -236,9 +236,8 @@ class $Mappr implements _i1.AutoMapprInterface {
         return null;
       }
       return (_map__i2$ValueHolderDto_To__i2$ValueHolder(
-            (model as _i2.ValueHolderDto?),
-          )
-          as TARGET);
+        (model as _i2.ValueHolderDto?),
+      ) as TARGET);
     }
     throw Exception('No ${model.runtimeType} -> $targetTypeOf mapping.');
   }